- Rename base.py to base_tool.py - Remove redundant ToolInputModel
