+++ atomic_agents/lib/tools/calculator_tool.py
@@ -3,7 +3,7 @@ from rich.console import Console
 from sympy import sympify
 
 from atomic_agents.agents.base_agent import BaseIOSchema
-from atomic_agents.lib.tools.base import BaseTool, BaseToolConfig
+from atomic_agents.lib.tools.base_tool import BaseTool, BaseToolConfig
 
 ################
 # INPUT SCHEMA #
