+++ atomic_agents/lib/tools/content_scraping_tool.py
@@ -9,7 +9,7 @@ from rich.console import Console
 from rich.markdown import Markdown
 
 from atomic_agents.agents.base_agent import BaseIOSchema
-from atomic_agents.lib.tools.base import BaseTool, BaseToolConfig
+from atomic_agents.lib.tools.base_tool import BaseTool, BaseToolConfig
 from atomic_agents.lib.utils.scraping.pdf_to_markdown import PdfToMarkdownConverter
 from atomic_agents.lib.utils.scraping.url_to_markdown import UrlToMarkdownConverter
