+++ atomic_agents/lib/tools/yelp_restaurant_finder_tool.py
@@ -7,7 +7,7 @@ from pydantic import Field
 from rich.console import Console
 
 from atomic_agents.agents.base_agent import BaseIOSchema
-from atomic_agents.lib.tools.base import BaseTool, BaseToolConfig
+from atomic_agents.lib.tools.base_tool import BaseTool, BaseToolConfig
 
 ################
 # INPUT SCHEMA #
