+++ atomic_agents/lib/tools/yt_transcript_scraper_tool.py
@@ -7,7 +7,7 @@ from rich.console import Console
 from youtube_transcript_api import NoTranscriptFound, TranscriptsDisabled, YouTubeTranscriptApi
 
 from atomic_agents.agents.base_agent import BaseIOSchema
-from atomic_agents.lib.tools.base import BaseTool, BaseToolConfig
+from atomic_agents.lib.tools.base_tool import BaseTool, BaseToolConfig
 
 ################
 # INPUT SCHEMA #
