+++ docs/api/atomic_agents/lib/tools/index.md
@@ -13,4 +13,4 @@
 - [CalculatorTool](./calculator_tool.md)
 - [ContentScrapingTool](./content_scraping_tool.md)
 - [Yelp Restaurant Finder Tool](./yelp_restaurant_finder_tool.md)
-- [Yt Transcript Scraper](./yt_transcript_scraper.md)
+- [Yt Transcript Scraper](./yt_transcript_scraper_tool.md)
