+++ tests/lib/tools/test_base.py
@@ -1,6 +1,6 @@
 import pytest
 from pydantic import BaseModel
-from atomic_agents.lib.tools.base import BaseToolConfig, BaseTool
+from atomic_agents.lib.tools.base_tool import BaseToolConfig, BaseTool
 from atomic_agents.agents.base_agent import BaseIOSchema
 
 # Mock classes for testing
