+++ tests/lib/tools/test_content_scraping_tool.py
@@ -6,7 +6,7 @@ from atomic_agents.lib.tools.content_scraping_tool import (
     ContentScrapingToolOutputSchema,
     ContentScrapingToolConfig,
 )
-from atomic_agents.lib.tools.base import BaseTool, BaseToolConfig
+from atomic_agents.lib.tools.base_tool import BaseTool, BaseToolConfig
 from atomic_agents.lib.utils.scraping.pdf_to_markdown import PdfToMarkdownConverter
 from atomic_agents.lib.utils.scraping.url_to_markdown import UrlToMarkdownConverter
